refactor(table): move overlay staleness checks into lance-table
Deciding which rows an overlay makes stale with respect to an index reads
only fragment and index metadata: the coverage bitmaps, the overlay
committed_version, and the indexed field ids. None of that needs the read
path, so it moves to lance-table alongside the overlay format itself.

lance::dataset::overlay keeps the read-resolution half and re-exports the
three functions its callers use.
